fix: compatibility with new fw / full error codes

+++ PostCodeSerialMonitor/Models/ErrorMaskDefinition.cs
@@ -1,3 +1,4 @@
+using System;
 using CsvHelper.Configuration.Attributes;
 
 namespace PostCodeSerialMonitor.Models;
@@ -12,10 +13,10 @@ public class ErrorMaskDefinition
     public CodeFlavor CodeFlavor { get; set; } = CodeFlavor.UNKNOWN;
 
     [TypeConverter(typeof(HexNumberConverter))]
-    public uint Bitmask { get; set; }
+    public UInt64 Bitmask { get; set; }
 
     [TypeConverter(typeof(HexNumberConverter))]
-    public uint Code { get; set; }
+    public UInt64 Code { get; set; }
 
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
