+++ PostCodeSerialMonitor/Models/PostCodeDefinition.cs
@@ -1,3 +1,4 @@
+using System;
 using CsvHelper.Configuration.Attributes;
 
 namespace PostCodeSerialMonitor.Models;
@@ -12,9 +13,9 @@ public class PostCodeDefinition
     public CodeFlavor CodeFlavor { get; set; } = CodeFlavor.UNKNOWN;
 
     [TypeConverter(typeof(HexNumberConverter))]
-    public uint Code { get; set; }
+    public UInt64 Code { get; set; }
     [TypeConverter(typeof(HexNumberConverter))]
-    public uint? Bitmask { get; set; }
+    public UInt64? Bitmask { get; set; }
     public bool IsError { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
